% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{Number of interations of binomial}

\item{n}{Number of trials}

\item{p}{Probability of success in each trial}
}
\value{
Number of successes in each simulation
}
\description{
mybin
}
