% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{myhyper}
\usage{
myhyper(iter, N, r, n)
}
\arguments{
\item{iter}{Number of iterations of simulation}

\item{N}{Total number of elements}

\item{r}{Number of successes in N elements}

\item{n}{Number of elements drawn}
}
\value{
A barplot of the hypergeometric simulation and a probability table
}
\description{
myhyper
}
