% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The upper bound of the lower tail}
}
\value{
A plot of the normal distribution with the lower tail shaded, the area pasted onto the plot, and a output of the probability.
}
\description{
myncurve
}
